Add logging functionality to save request and response data

+++ src/Log.php
@@ -23,40 +23,55 @@ final class Log
      */
     public function save(Request $request, Response $response): void
     {
-        $log = [
-            'request' => [
-                'date' => date('Y-m-d H:i:s'),
-                'method' => $request->getRequestMethod(),
-                'uri' => $request->getUri(),
-                'query_string' => $request->getQueryString(),
-                'header' => $request->getHeader(),
-                'get_params' => $request->getGetParams(),
-                'post_params' => $request->getPostParams(),
-                'middleware_params' => $request->getMiddlewareParams()
-            ],
-            'response' => [
-                'status_code' => $response->getStatusCode(),
-                'body' => $response->getBody(),
-                'redirection' => $response->getRedirection() ?
-                    $response->getRedirection()['url']
-                    : null,
-                'execution_time' => number_format(microtime(true) - KX_START, 4),
-            ],
-        ];
 
-        $log = json_encode($log, JSON_PRETTY_PRINT) . ',';
+        // log model class check
+        if (!class_exists('KX\Model\Logs')) {
 
-        $logFile = KX_ROOT . 'app/Storage/logs/' . $response->getStatusCode() . '_' .
-            date('Ymd') . '_' . Helper::getIp() . '.log';
+            $log = [
+                'request' => [
+                    'date' => date('Y-m-d H:i:s'),
+                    'method' => $request->getRequestMethod(),
+                    'uri' => $request->getUri(),
+                    'query_string' => $request->getQueryString(),
+                    'header' => $request->getHeader(),
+                    'get_params' => $request->getGetParams(),
+                    'post_params' => $request->getPostParams(),
+                    'middleware_params' => $request->getMiddlewareParams()
+                ],
+                'response' => [
+                    'status_code' => $response->getStatusCode(),
+                    'body' => $response->getBody(),
+                    'redirection' => $response->getRedirection() ?
+                        $response->getRedirection()['url']
+                        : null,
+                    'execution_time' => number_format(microtime(true) - KX_START, 4),
+                ],
+            ];
 
-        Helper::path('app/Storage/logs', true);
+            $log = json_encode($log, JSON_PRETTY_PRINT) . ',';
 
-        if (!file_exists($logFile)) {
-            touch($logFile);
+            $logFile = KX_ROOT . 'app/Storage/logs/' . $response->getStatusCode() . '_' .
+                date('Ymd') . '_' . Helper::getIp() . '.log';
+
+            Helper::path('app/Storage/logs', true);
+
+            if (!file_exists($logFile)) {
+                touch($logFile);
+            } else {
+                $log = PHP_EOL . $log;
+            }
+
+            file_put_contents($logFile, $log, FILE_APPEND);
         } else {
-            $log = PHP_EOL . $log;
+            $logModel = new \KX\Model\Logs();
+            $logModel->insert([
+                'endpoint' => $request->getUri(),
+                'status_code' => $response->getStatusCode(),
+                'auth_token' => Helper::authToken(),
+                'ip' => Helper::getIp(),
+                'header' => Helper::getUserAgent(),
+                'exec_time' => number_format(microtime(true) - KX_START, 4),
+            ]);
         }
-
-        file_put_contents($logFile, $log, FILE_APPEND);
     }
 }
